Suggested fix for Sentry issue: TypeError: array_key_exists(): Argument #2 () must be of type array, null given
Root cause: The SQL query on line 143 only selects 'externalid' column, but line 150 tries to use ['id'] to construct a CommunityEvent object. This results in ['id'] being null/undefined, causing the CommunityEvent constructor to not fetch any data (id is null), leaving ->communityevent as NULL. When setPrivate() is then called on line 151, it tries to call array_key_exists() on the NULL array.
Fix: Add 'communityevents.id' to the SELECT clause in the query so that ['id'] is available when constructing the CommunityEvent object

Confidence: high
Fix type: simple

Sentry issue: https://freegle.sentry.io/issues/6975846513/